Adjusted api to accept Cross-domain calls. Adjusted setPrice method of Room. Created initial website with list rooms page

+++ src/com/hotelreservation/model/Room.java
@@ -41,7 +41,7 @@ public class Room {
 		return price;
 	}
 
-	public void setPrice(int prc) {
+	public void setPrice(double prc) {
 		this.price = prc;
 	}
 	public String getDescription() {
